server/requirements.txt
fastapi
uvicorn[standard]
asyncpg
prisma
python-jose[cryptography]
passlib
opencv-python
pyzbar
python-multipart
Pillow
requests